Replaced with larger hand compressed gif
Hand done transparency compression resulted in better image quality than the 60mb original, and 5MB of size. couldnt find otption for transparency compression, wasn't doing it right, or it needed to be done in ffmpeg which seemed to have applied some dither or error-difference method of compression resulting in static parts of the image needlessly changing to lower quality with a diffused look.